Try adding values back in...
Adding it in the same repo is (I think) only to ensure the service is
restarted on secret change. These secrets won't change often.

And, if the secret is in that repository it won't be (shouldn't be?)
decrypted, but if it's in this one the cluster kustomization-secrets
should decrypt it. That's the theory!